docker_* helpers: pass usb host controllers to docker only if usb devices are connected, unify, bugfixes

+++ docker_repro.sh
@@ -5,8 +5,8 @@ DOCKER_IMAGE=$(grep -oP '^\s*-?\s*image:\s*\K(tlaurion/heads-dev-env:[^\s]+)' .c
 
 # Check if the Docker image was found
 if [ -z "$DOCKER_IMAGE" ]; then
-    echo "Error: Docker image not found in .circleci/config.yml"
-    exit 1
+	echo "Error: Docker image not found in .circleci/config.yml"
+	exit 1
 fi
 
 # Inform the user about the versioned CircleCI Docker image being used
@@ -14,23 +14,23 @@ echo "Using CircleCI Docker image: $DOCKER_IMAGE"
 
 # Function to display usage information
 usage() {
-    echo "Usage: $0 [OPTIONS] -- [COMMAND]"
-    echo "Options:"
-    echo "  CPUS=N  Set the number of CPUs"
-    echo "  V=1     Enable verbose mode"
-    echo "Command:"
-    echo "  The command to run inside the Docker container, e.g., make BOARD=BOARD_NAME"
+	echo "Usage: $0 [OPTIONS] -- [COMMAND]"
+	echo "Options:"
+	echo "  CPUS=N  Set the number of CPUs"
+	echo "  V=1     Enable verbose mode"
+	echo "Command:"
+	echo "  The command to run inside the Docker container, e.g., make BOARD=BOARD_NAME"
 }
 
 # Function to kill GPG toolstack related processes using USB devices
 kill_usb_processes() {
-    echo "Killing any GPG toolstack related processes on host currently using USB devices..."
-    sudo lsof /dev/bus/usb/00*/0* 2>/dev/null | awk 'NR>1 {print $2}' | xargs -r ps -p | grep -E 'scdaemon|pcscd' | awk '{print $1}' | xargs -r sudo kill -9
-    if [ $? -ne 0 ]; then
-        echo "Failed to kill GPG toolstack related processes using USB devices. Please run the following command manually:"
-        echo "sudo lsof /dev/bus/usb/00*/0* | awk 'NR>1 {print \$2}' | xargs -r ps -p | grep -E 'scdaemon|pcscd' | awk '{print \$1}' | xargs -r sudo kill -9"
-        exit 1
-    fi
+	# check if scdaemon or pcscd processes are using USB devices
+	if [ -d /dev/bus/usb ]; then
+		if sudo lsof /dev/bus/usb/00*/0* 2>/dev/null | awk 'NR>1 {print $2}' | xargs -r ps -p | grep -E 'scdaemon|pcscd' >/dev/null; then
+			echo "Killing GPG toolstack related processes using USB devices..."
+			sudo lsof /dev/bus/usb/00*/0* 2>/dev/null | awk 'NR>1 {print $2}' | xargs -r ps -p | grep -E 'scdaemon|pcscd' | awk '{print $1}' | xargs -r sudo kill -9
+		fi
+	fi
 }
 
 # Handle Ctrl-C (SIGINT) to exit gracefully
@@ -38,10 +38,10 @@ trap "echo 'Script interrupted. Exiting...'; exit 1" SIGINT
 
 # Check if --help or -h is provided
 for arg in "$@"; do
-    if [[ "$arg" == "--help" || "$arg" == "-h" ]]; then
-        usage
-        exit 0
-    fi
+	if [[ "$arg" == "--help" || "$arg" == "-h" ]]; then
+		usage
+		exit 0
+	fi
 done
 
 # Kill processes using USB devices
@@ -51,8 +51,16 @@ kill_usb_processes
 echo "----"
 echo "Usage reminder: The minimal command is 'make BOARD=XYZ', where additional options, including 'V=1' or 'CPUS=N' are optional."
 echo "For more advanced QEMU testing options, refer to targets/qemu.md and boards/qemu-*/*.config."
+echo
+echo "Type exit within docker image to get back to host if launched interactively!"
 echo "----"
-echo "Entering the Docker container. Type 'exit' to return to the host shell."
+echo
 
 # Execute the docker run command with the provided parameters
-docker run --device=/dev/bus/usb:/dev/bus/usb -e DISPLAY=$DISPLAY --network host --rm -ti -v $(pwd):$(pwd) -w $(pwd) $DOCKER_IMAGE -- "$@"
+if [ -d "/dev/bus/usb" ]; then
+	echo "--->Launching container with access to host's USB buses (some USB devices were connected to host)..."
+	docker run --device=/dev/bus/usb:/dev/bus/usb -e DISPLAY=$DISPLAY --network host --rm -ti -v $(pwd):$(pwd) -w $(pwd) $DOCKER_IMAGE -- "$@"
+else
+	echo "--->Launching container without access to host's USB buses (no USB devices was connected to host)..."
+	docker run -e DISPLAY=$DISPLAY --network host --rm -ti -v $(pwd):$(pwd) -w $(pwd) $DOCKER_IMAGE -- "$@"
+fi
